fix extract package and add read method
In the older code the extract method was reading the buffer and alocate the data in a variable.
In this version, we set the extract package to only check the buffer.

We also add a new method named read, which is responsible to receive a string and add the data from the buffer to this variable.
In this way, you can access the value of the sensor from this variable.

This version needs to be tested, check the node file to see how to use the method read().

Is very important for next updates that you change the format (string to float64), to publish directly the sensor value.

In the future, an alternative way is doing this using services. Mainly because you need to send a command to receive the response, this is a classic example of using services. Please try this!